requirements.txt
# Core dependencies
Flask==2.3.3                    # Web framework
Flask-Login==0.6.2              # Authentication management
Werkzeug==2.3.7                 # Used by Flask for various utilities
python-dotenv==1.0.0            # Environment variable management
bcrypt==4.0.1                   # Password hashing
schedule==1.2.0                 # Task scheduling
gunicorn==21.2.0                # WSGI HTTP server (for production)

# LinkedIn scraping
linkedin-jobs-scraper   # LinkedIn job scraping library

# LLM Integration
openai                  # OpenAI API client
tiktoken                # Token counting for OpenAI models

# Frontend
Bootstrap-Flask==2.3.0          # Bootstrap integration for Flask
Flask-WTF==1.1.1                # Form handling and CSRF protection
WTForms==3.0.1                  # Form validation and rendering

# Data processing
pandas==2.1.0                   # Data manipulation
beautifulsoup4==4.12.2          # HTML parsing for fallback scraping

# Testing
pytest==7.4.0                   # Testing framework
pytest-flask==1.2.0             # Flask-specific testing utilities

# Development tools
pylint==2.17.5                  # Code linting
black==23.7.0                   # Code formatting
